requirements.txt
numpy>=1.24.2
polars>=0.20.0
tqdm>=4.66.0
pytest
pyhamcrest